LLM-Enhanced-Query-Interface-Demand-forecasting/requirements.txt
langchain
langchain[anthropic]
flask
python-dotenv
requests
SQLAlchemy
sqlparse
google.generativeai
sqlite3
re
